added COORD2WELL, updated hist-plotting code
added the function coord2well--this was previously in utility and is
required for some function in this package.

updated plothistdensity and plothistseries to refer to the horizontal
axis as 'xgrid' rather than 'tgrid', since these functions can be used
to plot any series of histograms, not just those in time series.

wrote a comment to say what fcsseg1d.m does.